+++ app/api/core/infrastructure/express/template/GetTemplatesController.ts
@@ -1,5 +1,5 @@
 import { AbstractController } from 'api/common.v2/infrastructure/AbstractController';
-import { TemplateDBO } from 'api/core/infrastructure/mongodb/template/schemas/TemplateDBO';
+import { TemplateDBO } from 'api/core/infrastructure/mongodb/template/DBOs/TemplateDBO';
 import { getConnection } from 'api/common.v2/database/getConnectionForCurrentTenant';
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
