+++ app/api/core/infrastructure/mongodb/template/TemplateMappers.ts
@@ -8,8 +8,8 @@ import { RelationshipProperty } from '../../../domain/template/RelationshipPrope
 import { Template } from '../../../domain/template/Template';
 import { V1RelationshipProperty } from '../../../domain/template/V1RelationshipProperty';
 import { mapPropertyQuery } from './QueryMapper';
-import { TraverseQueryDBO } from './schemas/RelationshipsQueryDBO';
-import { RelationshipPropertyDBO, TemplateDBO } from './schemas/TemplateDBO';
+import { TraverseQueryDBO } from './DBOs/RelationshipsQueryDBO';
+import { RelationshipPropertyDBO, TemplateDBO } from './DBOs/TemplateDBO';
 
 type PropertyDBO = TemplateDBO['properties'][number];
 
@@ -17,6 +17,7 @@ function propertyToApp(
   property: RelationshipPropertyDBO,
   _templateId: TemplateDBO['_id']
 ): RelationshipProperty;
+//@ts-ignore
 function propertyToApp(property: PropertySchema, _templateId: TemplateDBO['_id']): Property;
 function propertyToApp(property: PropertyDBO, _templateId: TemplateDBO['_id']): Property {
   const templateId = MongoIdHandler.mapToApp(_templateId);
